In main.py:
Using Yfinace got 365 days of data for whatever stock is hard coded into the code.

Using matplotlib graphed the 365 days worth of data.
Added labels like title, ylabel, xlabel and added a basic legend to the graph.

Using np.arange made a trendline for the years worth of data, trendline is based on the average price. Also made the trendline green if the trendline was positive and red if the trendline was negitive.